requirements.txt
Django==3.2.10
djangorestframework==3.12.4
pytest==6.2.5
pytest-django==4.5.2
coverage==5.5
